requirements.txt
discord.py>=2.2.2
peewee>=3.16.0
humanize>=4.6.0